Final Project Submission
This is the last upload to the project which is fully executable and should compile correctly. All assets, files, scripts and necessary files to run the application are held here.

If an API error appears mentioning "Requires API level 26+ please upgrade" ignore this message. Close Unity, reopen the project and try to build the project again. This is just a consistent problem with the ARCore SDK.